requirements.txt
torch==2.1.0
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0
hydra-core==1.3.2
omegaconf==2.3.0
tabulate==0.9.0